requirements.txt
gallery-dl
Pillow
requests
yt-dlp
